class browser implementation
