+++ components/inspectit-ocelot-configurationserver-ui/src/components/common/class-browser/ClassBrowserDialog.js
@@ -5,31 +5,23 @@ import { Dialog } from 'primereact/dialog';
 import { Dropdown } from 'primereact/dropdown';
 import { InputText } from 'primereact/inputtext';
 import { Tree } from 'primereact/tree';
-import { RadioButton } from 'primereact/radiobutton';
+import CBTableNode from './CBTableNode';
 import _ from 'lodash';
+import { transformClassStructureToTableModel } from './ClassBrowserUtilities';
 
 /**
  * The class browser dialog.
  */
-const ClassBrowserDialog = ({ visible, onHide, onApply }) => {
+const ClassBrowserDialog = ({ visible, onHide, onSelect }) => {
   const [selectedAgent, setSelectedAgent] = useState(null);
   const [selectedMethod, setSelectedMethod] = useState(null);
-  const [searchQuery, setSearchQuery] = useState("");
+  const [searchQuery, setSearchQuery] = useState('');
   const [treeModel, setTreeModel] = useState([]);
-
-  const agents = [
-    {
-      metaInformation: {
-        agentId: 'sample-agent',
-      },
-      attributes: { service: 'service-name' },
-    },
-  ];
 
   const classStructure = [
     { name: 'java.lang.String', type: 'class', methods: ['public String toString()', 'public String equals(java.lang.Object)'] },
     { name: 'java.lang.Runnable', type: 'interface', methods: ['public void run()'] },
-    { name: 'org.lang.String', type: 'class', methods: ['public String toString()', 'public String equals(java.lang.Object)'] },
+    { name: 'org.lang.String', type: 'class', methods: ['public String toString()', 'public String equals(java.lang.Object) public String equals(java.lang.Object)'] },
   ];
 
   const existingAgent = [
@@ -38,163 +30,24 @@ const ClassBrowserDialog = ({ visible, onHide, onApply }) => {
     { label: 'InspectIT Agent (31360@NB171217MO)', value: '31360@NB171217MO' },
   ];
 
+  // generate table model
   useEffect(() => {
     const result = [];
-
-    _.each(classStructure, ({ name, type, methods }) => {
-      const packages = name.split('.');
-      const className = packages.pop();
-
-      // find target package
-      let targetNode = result;
-      let packagePath;
-      packages.forEach((packageName) => {
-        let target;
-        if (targetNode === result) {
-          target = _.find(targetNode, { label: packageName });
-        } else {
-          target = _.find(targetNode.children, { label: packageName });
-        }
-
-        if (packagePath) {
-          packagePath += '.' + packageName;
-        } else {
-          packagePath = packageName;
-        }
-
-        if (!target) {
-          target = {
-            key: packagePath,
-            label: packageName,
-            type: 'package',
-            children: [],
-            selectable: false,
-          };
-
-          if (targetNode === result) {
-            result.push(target);
-          } else {
-            targetNode.children.push(target);
-          }
-        }
-
-        targetNode = target;
-      });
-
-      // add type
-      const typeNode = {
-        key: name,
-        label: className,
-        type,
-        children: [],
-        selectable: false,
-      };
-      targetNode.children.push(typeNode);
-
-      // add methods
-      methods.forEach((method) => {
-        typeNode.children.push({
-          key: name + ': ' + method,
-          label: method,
-          type: 'method',
-        });
-      });
-      console.log(result);
-    });
-
+    _.each(classStructure, (classElement) => transformClassStructureToTableModel(result, classElement));
     setTreeModel(result);
-    console.log(result);
   }, []);
-
-  const nodes = [
-    {
-      key: 'java',
-      label: 'java',
-      type: 'package',
-      children: [
-        { key: '0-0', label: 'Getting Started', url: 'https://reactjs.org/docs/getting-started.html' },
-        { key: '0-1', label: 'Add React', url: 'https://reactjs.org/docs/add-react-to-a-website.html' },
-        { key: '0-2', label: 'Create an App', url: 'https://reactjs.org/docs/create-a-new-react-app.html' },
-        { key: '0-3', label: 'CDN Links', url: 'https://reactjs.org/docs/cdn-links.html' },
-      ],
-    },
-  ];
-
-  const nodeTemplate = ({ label, type, key }) => {
-    if (type === 'package') {
-      return (
-        <>
-          <style jsx>{`
-            i {
-              margin-right: 0.5rem;
-              color: gray;
-            }
-            span {
-              font-family: monospace;
-            }
-          `}</style>
-          <i className="pi pi-folder"></i> <span>{label}</span>
-        </>
-      );
-    } else {
-      let typeIcon;
-      let typeClass;
-      let selectionButton;
-      if (type === 'class') {
-        typeIcon = 'c';
-        typeClass = 'theme-class';
-      } else if (type === 'interface') {
-        typeIcon = 'i';
-        typeClass = 'theme-interface';
-      } else if (type === 'method') {
-        typeIcon = 'm';
-        typeClass = 'theme-method';
-
-        selectionButton = <RadioButton value={key} onChange={(e) => setSelectedMethod(e.value)} checked={selectedMethod === key} />;
-      }
-
-      return (
-        <>
-          <style jsx>{`
-            .theme-class {
-              background-color: #719cbb;
-            }
-            .theme-interface {
-              background-color: #55ab73;
-            }
-            .theme-method {
-              background-color: #c855f7;
-            }
-            .type-icon {
-              border-radius: 50%;
-              width: 1rem;
-              height: 1rem;
-              text-align: center;
-              color: white;
-              font-size: 0.75rem;
-              margin-right: 0.5rem;
-              font-family: monospace;
-              line-height: 1rem;
-            }
-            span {
-              flex-grow: 1;
-            }
-          `}</style>
-          <div className={'type-icon ' + typeClass}>{typeIcon}</div>
-          <span>{label}</span>
-          {selectionButton}
-        </>
-      );
-    }
-  };
 
   // the dialogs footer
   const footer = (
     <div>
-      <Button label="Select" onClick={() => onApply()} />
+      <Button label="Select" onClick={() => onSelect()} disabled={!selectedMethod} />
       <Button label="Cancel" className="p-button-secondary" onClick={onHide} />
     </div>
   );
+
+  const tableNodeTemplate = ({ label, type, key }) => {
+    return <CBTableNode label={label} type={type} value={key} onChange={setSelectedMethod} selectedMethod={selectedMethod} />;
+  };
 
   return (
     <>
@@ -215,17 +68,19 @@ const ClassBrowserDialog = ({ visible, onHide, onApply }) => {
 
         .content :global(.p-tree) {
           width: 100%;
-          height: 20rem;
           margin-top: 0.5rem;
         }
-
+        .content :global(.p-tree-container) {
+          height: 20rem;
+        }
         .content :global(.p-treenode-content) {
           display: flex;
         }
-        .content :global(.p-treenode-label) {
+        .content :global(.p-tree .p-tree-container .p-treenode .p-treenode-content .p-treenode-label) {
           flex-grow: 1;
           align-items: center;
           display: flex;
+          overflow: hidden;
         }
       `}</style>
 
@@ -253,7 +108,7 @@ const ClassBrowserDialog = ({ visible, onHide, onApply }) => {
             <Button label="Search" />
           </div>
 
-          <Tree value={treeModel} nodeTemplate={nodeTemplate} />
+          <Tree value={treeModel} nodeTemplate={tableNodeTemplate} />
         </div>
       </Dialog>
     </>
